bug fixed and rename files
